refactor : 불필요한 BaseTableViewCell 제거
